+++ app/services/pipelines/create.rb
@@ -22,7 +22,8 @@ module Pipelines
     private
 
     def post_save
-      r_mkdir(@pipeline.working_dir)
+      r_mkdir(@pipeline.inputs_dir)
+      r_mkdir(@pipeline.outputs_dir)
       create_computations
     end